unicode characters work for genes and ciphers

+++ eye/src/cipher/alberti.rs
@@ -2,12 +2,12 @@ use crate::errors::*;
 use crate::cipher::{Cipher, Wheel};
 
 pub struct Alberti {
-    plainwheel: Wheel<u32>,
-    cipherwheel: Wheel<u32>
+    plainwheel: Wheel<u16>,
+    cipherwheel: Wheel<u16>
 }
 
 impl Alberti {
-    pub fn new(plainwheel: Wheel<u32>, cipherwheel: Wheel<u32>) -> Self {
+    pub fn new(plainwheel: Wheel<u16>, cipherwheel: Wheel<u16>) -> Self {
         Alberti {
             plainwheel,
             cipherwheel
@@ -16,8 +16,8 @@ impl Alberti {
 }
 
 impl Cipher for Alberti {
-    fn encode(&mut self, message: &[u32]) -> Result<Vec<u32>> {
-        let mut ciphertext: Vec<u32> = vec![];
+    fn encode(&mut self, message: &[u16]) -> Result<Vec<u16>> {
+        let mut ciphertext: Vec<u16> = vec![];
         for letter in message {
             ciphertext.push( match self.plainwheel.iter()
                 .position(|x| x == letter) {
@@ -29,8 +29,8 @@ impl Cipher for Alberti {
 
         Ok(ciphertext)
     }
-    fn decode(&mut self, ciphertext: &[u32]) -> Result<Vec<u32>> {
-        let mut plaintext: Vec<u32> = vec![];
+    fn decode(&mut self, ciphertext: &[u16]) -> Result<Vec<u16>> {
+        let mut plaintext: Vec<u16> = vec![];
         for letter in ciphertext {
             plaintext.push( match self.cipherwheel.iter()
                 .position(|x| x == letter) {
