+++ eye/src/cipher/mod.rs
@@ -7,6 +7,6 @@ pub use substitution::Substitution;
 pub use std::collections::VecDeque as Wheel;
 
 pub trait Cipher {
-    fn encode(&mut self, message: &[u32]) -> Result<Vec<u32>>;
-    fn decode(&mut self, ciphertext: &[u32]) -> Result<Vec<u32>>;
+    fn encode(&mut self, message: &[u16]) -> Result<Vec<u16>>;
+    fn decode(&mut self, ciphertext: &[u16]) -> Result<Vec<u16>>;
 }
